API/requirements.txt
Flask==2.1.2
kafka-python==2.0.2
requests==2.22.0
Pillow==7.0.0
flask_socketio==5.2.0